# swagger.yaml
openapi: 3.0.0
info:
  title: InstaCode API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/user/login:
    post:
      summary: Authenticate user credentials and return a JWT & User
      operationId: authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: 'minicodelab@gmail.com'
                password:
                  type: string
                  example: 'MiniCodeLab@2022'
              required:
                - email
                - password
      responses:
        '200':
          description: User details, a JWT access token and a refresh token cookie
          headers:
            Set-Cookie:
              description: ''
              schema:
                type: string
                example:
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: '5eb12e197e06a76ccdefc121'
                  username:
                    type: string
                    example: 'MiniCodeLab'
                  email:
                    type: string
                    example: 'minicodelab@gmail.com'
                  emoji:
                    type: string
                    example: '🙈'
                  jwtToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZWIxMmUxOTdlMDZhNzZjY2RlZmMxMjEiLCJpZCI6IjVlYjEyZTE5N2UwNmE3NmNjZGVmYzEyMSIsImlhdCI6MTU4ODc1ODE1N30.xR9H0STbFOpSkuGA9jHNZOJ6eS7umHHqKRhI807YT1Y'
                  codes:
                    type: array
                    example: ['']
                  favCodes:
                    type: array
                    example: ['']
        '400':
          description: The email or password is incorrect
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Email or password is incorrect'
  /users:
    get:
      summary: Get a list of all users
      description: Restricted to admin users.
      operationId: getAllUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: An array of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: '5eb12e197e06a76ccdefc121'
                    firstName:
                      type: string
                      example: 'MiniCodeLab'
                    lastName:
                      type: string
                      example: 'MiniCodeLab'
                    username:
                      type: string
                      example: 'MiniCodeLab'
                    role:
                      type: string
                      example: 'Admin'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/{id}:
    parameters:
      - in: path
        name: id
        description: User id
        required: true
        example: '5eb12e197e06a76ccdefc121'
        schema:
          type: string
    get:
      summary: Get a single user by id
      description: Admin users can access any user record, regular users are restricted to their own user record.
      operationId: getUserById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Details of the specified user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: '5eb12e197e06a76ccdefc121'
                  firstName:
                    type: string
                    example: 'MiniCodeLab'
                  lastName:
                    type: string
                    example: 'MiniCodeLab'
                  username:
                    type: string
                    example: 'MiniCodeLab'
                  role:
                    type: string
                    example: 'Admin'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Unauthorized'
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Not Found'